backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
python-multipart>=0.0.9,<1.0.0
Pillow>=12.0.0,<13.0.0
pytesseract>=0.3.13,<1.0.0
icalendar>=6.3.2,<7.0.0
python-dotenv>=1.0.0,<2.0.0
pdf2image>=1.16.3,<2.0.0
openpyxl>=3.1.2,<4.0.0
asyncpg>=0.29.0,<1.0.0
SQLAlchemy>=2.0.23,<3.0.0
alembic>=1.13.0,<2.0.0
pydantic>=2.5.2,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
azure-storage-blob>=12.19.0,<13.0.0
azure-identity>=1.15.0,<2.0.0
azure-keyvault-secrets>=4.7.0,<5.0.0
stripe>=7.7.0,<10.0.0
slowapi>=0.1.9,<1.0.0
python-magic>=0.4.27,<1.0.0
openai>=1.54.0,<2.0.0
tenacity>=8.2.0,<10.0.0

nh3>=0.2.14,<1.0.0
sentry-sdk[fastapi]>=2.0.0,<3.0.0
